+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: chideyuk <chideyuk@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/10 13:33:37 by chideyuk          #+#    #+#             */
-/*   Updated: 2022/01/19 19:49:04 by chideyuk         ###   ########.fr       */
+/*   Updated: 2022/01/20 17:20:01 by chideyuk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ int		ft_strncmp(const char	*s1, const char	*s2, size_t	len);
 void	*ft_memchr(const void	*b, int	c, size_t	len);
 int		ft_memcmp(const void	*b1, const void	*b2, size_t	len);
 char	*ft_strnstr(const char	*big, const char	*little, size_t	len);
-int		ft_atoi(const char	*nptr);
+long	ft_atoi(const char	*nptr);
 void	*ft_calloc(size_t	number, size_t	size);
 char	*ft_strdup(const char	*str);
 char	*ft_substr(char const	*s, unsigned int	start, size_t	len);
